requirements.txt

math
mathplotlib
mesh
numpy
numpy-stl
scipy
